Finish refactoring write page
